LayoutTests/platform/chromium-mac/fast/js/modify-non-references-expected.txt

PASS function f() { g()++; } f.toString() is 'function f() { g()++; }'
PASS function f() { g()--; } f.toString() is 'function f() { g()--; }'
PASS function f() { ++g(); } f.toString() is 'function f() { ++g(); }'
PASS function f() { --g(); } f.toString() is 'function f() { --g(); }'
PASS function f() { g() = 1; } f.toString() is 'function f() { g() = 1; }'
PASS function f() { g() += 1; } f.toString() is 'function f() { g() += 1; }'
FAIL g()++ should throw ReferenceError: Postfix ++ operator applied to value that is not a reference.. Threw exception ReferenceError: Invalid left-hand side expression in postfix operation.
FAIL g()-- should throw ReferenceError: Postfix -- operator applied to value that is not a reference.. Threw exception ReferenceError: Invalid left-hand side expression in postfix operation.
FAIL ++g() should throw ReferenceError: Prefix ++ operator applied to value that is not a reference.. Threw exception ReferenceError: Invalid left-hand side expression in prefix operation.
FAIL --g() should throw ReferenceError: Prefix -- operator applied to value that is not a reference.. Threw exception ReferenceError: Invalid left-hand side expression in prefix operation.
FAIL g() = 1 should throw ReferenceError: Left side of assignment is not a reference.. Threw exception ReferenceError: Invalid left-hand side in assignment.
FAIL g() += 1 should throw ReferenceError: Left side of assignment is not a reference.. Threw exception ReferenceError: Invalid left-hand side in assignment.
PASS successfullyParsed is true

TEST COMPLETE
